requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Strands AI Framework
strands-agents>=1.0.0
strands-agents-tools>=0.2.9
strands-agents-builder>=0.1.10

# AWS & Bedrock
boto3>=1.34.90
botocore>=1.34.90

# Git Operations
GitPython==3.1.41

# HTTP & API
requests==2.31.0
urllib3>=1.25.4,<2.1

# Data Processing
PyYAML==6.0.1
jsonschema==4.20.0

# Database
# SQLite is included in Python standard library (no package needed)

# Logging & Monitoring
colorlog==6.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0

ruamel.yaml>=0.17.0
